requirements.txt
pandas==1.4.4
plotly==5.10.0
snowflake_snowpark_python==0.9.0
pyarrow<8.1.0,>=8.0.0

